// src/types/Monitoring.types.d.ts
export interface thresholdType {
  vibration?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  noise?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
  temperature?: {
    red?: number;
    yellow?: number;
    green?: number;
  };
}
// ...
export interface notificationType {
  enable: boolean;
  channels: [
    {
      name: string;
      enable: boolean;
    },
    {
      name: string;
      enable: boolean;
    },
    {
      name: string;
      enable: boolean;
    },
  ];
}

export interface configurationDetailsType {
  notifications: notificationType;
  thresholds: thresholdType;
}